Move repositories and data initialzer
