+++ backend/src/main/java/com/videolibrary/backend/domain/service/SeriesService.java
@@ -1,7 +1,7 @@
 package com.videolibrary.backend.domain.service;
 
 import com.videolibrary.backend.domain.entity.Series;
-import com.videolibrary.backend.infrastructure.repository.SeriesRepository;
+import com.videolibrary.backend.infrastructure.sql.repository.SeriesRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
