+++ backend/src/main/java/com/videolibrary/backend/infrastructure/sql/repository/EpisodeRepository.java
@@ -1,4 +1,4 @@
-package com.videolibrary.backend.infrastructure.repository;
+package com.videolibrary.backend.infrastructure.sql.repository;
 
 import com.videolibrary.backend.domain.entity.Episode;
 import org.springframework.data.jpa.repository.JpaRepository;
